Edited for grammar, punctuation, etc.
line 2244: is "consume" the word you want to use here? I'm not sure if it has a special meaning for this context.

line 2456: "on terms of them" - is this correct? I'm not sure what it means.

line 2530: "of the possible state to select" - again, I'm not sure what this should communicate.