Edited for grammar, punctuation, etc.
line 10: What is "it" in this case? Is it the store or the record? It might be helpful to clear that up so the reader knows what the subject is.

line 179: Is this correct? "It will return null if the record is available, or null otherwise."

lines 209-211: Check this sentence to make sure that I have edited it according to your intended meaning: "Suppose we are in
the context of a friend and we know they have an **articles** property
that represents all the articles belonging to another friend."

line 235: I'm not sure whether you want to say that you specify the ids on the payload ("With the first strategy, we'll specify the ids on the payload...") or "With the first strategy on the payload, we'll specify the ids..."

line 520: Is "It" the promise array? You might want to replace "It" with the name of your subject.